DAPRH/requirements.txt
certifi @ file:///croot/certifi_1665076670883/work/certifi
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
faiss-gpu==1.6.3
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
h5py==3.7.0
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.6.0
matplotlib==3.6.2
metric-learn==0.6.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
numpy @ file:///croot/numpy_and_numpy_base_1668593735768/work
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scikit-learn==1.2.0
scipy==1.9.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
termcolor==2.1.1
threadpoolctl==3.1.0
torch==1.8.0
torchaudio==0.8.0a0+a751e1d
torchvision==0.9.0
tqdm==4.64.1
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
yacs==0.1.8
